Detects Duplicate AccountTypes
The basis of this code is to 'detect' duplicate accountTypes i.e., when the user tries to make a Main account when they already have one.

For the time being I have set it so that making a duplicate accountType is not possible, however, this can be easily manipulated, for example, within the if-statement in newAccountCreation you can create a feature which will allow a duplicate account and rename the accountType.

How it works:
AccountType was added in the Account.java constructor - meaning you now have to define what type of account you are making (OOP).
I made an arraylist called accountTypes in Customer.java and a getter simply for easy retrieval.