Pay Feature
This feature is based on the duplicateAccount detection.

The for-loops in Customer.java loop through the account of the customer and make the payment when they reach the desired account.

The if-statements in NewBank.java get user input to determine which account they would like to pay from and pay to - this data is passed to the aforementioned for-loops.

I created a new function in NewBankClientHandler.java called getUserInput() - this allows you to get user input from another class.

Minor change - also added an if-statement making sure that the amount they want to pay is not equal to 0 i.e., they can't pay someone £0